requirements.txt
flask
groq==0.9.0
deep-translator
python-dotenv
gunicorn